assembly: fix vost and cinost init scripts
Init scripts were calling status_of_proc with an empty string "" as
first parameter ($DAEMON), which disappeared when passed on to
pidofproc. While this was previously ignored, lsb-base version 4.1
returns an error if the parameter is missing. This fixes the error by
fixing parameter quotes and passing the (non-empty) value of $DAEMON.